guacamole: define Soundtuner packages/activity
Add package name strings and the string for the Soundtuner activity in the overlay for PartsBin.

Those are the default values, which are present in res/deviceparts too, but if the DAX implentation is changed to other files, these strings need to be adapted to make the Soundtuner menu in PartsBin become visible.

First strings make the on/off switch for the Dolby service become visible.
Second and third strings make the option to get to the configuration UI for Dolby appear.

Defining none of the strings with a non-OOS package or service, will result in PartsBin NOT displaying the Sound menu, as the installed packages are checked by PartsBin based on the strings mentioned above.

Change-Id: Id6d5784bdff3fed8f2427660366c49c76ec9554d