guacamole: [test] adapt EU build fingerprint
As we are using the build fingerprint for global devices and "fool" the rest of the fingerprint, we should at least keep in mind that the EU version using a slightly bit different build fingerprint.
The device variant's name end with "_EEA", so we try and inject that in on the init.

Change-Id: Ida3ea9063efa77ca476443015474f33270f1a032